- Move most of `ProcessTools` to `ProcessUtilities` and make public.
Easier to debug and diagnose via PowerShell, especially for end users.
